# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: COMET LKW-Verladungsverwaltung API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shipments:
    get:
      operationId: listShipments
      tags: [shipments]
      summary: List shipments (Verladungen)
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
        - name: speditionId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: lkwArt
          in: query
          schema:
            type: ["string", "null"]
        - name: relation
          in: query
          schema:
            type: ["string", "null"]
        - name: kennzeichen
          in: query
          schema:
            type: ["string", "null"]
        - name: tor
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Shipments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shipment"
    post:
      operationId: createShipment
      tags: [shipments]
      summary: Create shipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShipmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
  /shipments/bulk:
    post:
      operationId: bulkCreateShipments
      tags: [shipments]
      summary: Bulk create shipments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkShipmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shipment"
  /shipments/{id}:
    get:
      operationId: getShipment
      tags: [shipments]
      summary: Get shipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
    patch:
      operationId: updateShipment
      tags: [shipments]
      summary: Update shipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShipmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
    delete:
      operationId: deleteShipment
      tags: [shipments]
      summary: Delete shipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /shipments/{id}/lock:
    post:
      operationId: lockShipment
      tags: [shipments]
      summary: Lock shipment for spedition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
  /shipments/{id}/unlock:
    post:
      operationId: unlockShipment
      tags: [shipments]
      summary: Unlock shipment for spedition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unlocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
  /shipments/{id}/history:
    get:
      operationId: getShipmentHistory
      tags: [shipments]
      summary: Get shipment change history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: History
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEntry"
...
components:
  schemas:
...
    Shipment:
      type: object
      required: [id, status, createdAt, updatedAt, cometBearbeitet, gesperrtFuerSpedition]
      properties:
        id:
          type: integer
        bezeichnung:
          type: ["string", "null"]
        kennzeichen:
          type: ["string", "null"]
        relation:
          type: ["string", "null"]
        speditionId:
          type: ["integer", "null"]
        speditionName:
          type: ["string", "null"]
        subSpeditionId:
          type: ["integer", "null"]
        subSpeditionName:
          type: ["string", "null"]
        bemerkungen:
          type: ["string", "null"]
        telefon:
          type: ["string", "null"]
        etaDate:
          type: ["string", "null"]
        etaTime:
          type: ["string", "null"]
        ataDate:
          type: ["string", "null"]
        ataTime:
          type: ["string", "null"]
        lkwArt:
          type: ["string", "null"]
          enum: [Container, Anlieferung, Abholung, Retoure, Sonstiges, null]
        status:
          type: string
          enum: [Angemeldet, Erwartet, Angekommen, Verladen, Abgefertigt, Storniert]
        tor:
          type: ["string", "null"]
        cometBearbeitet:
          type: boolean
        gesperrtFuerSpedition:
          type: boolean
        createdBy:
          type: ["integer", "null"]
        createdByName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedBy:
          type: ["integer", "null"]
        updatedByName:
          type: ["string", "null"]
        updatedAt:
          type: string
    ShipmentInput:
      type: object
      properties:
        bezeichnung:
          type: ["string", "null"]
        kennzeichen:
          type: ["string", "null"]
        relation:
          type: ["string", "null"]
        speditionId:
          type: ["integer", "null"]
        subSpeditionId:
          type: ["integer", "null"]
        bemerkungen:
          type: ["string", "null"]
        telefon:
          type: ["string", "null"]
        etaDate:
          type: ["string", "null"]
        etaTime:
          type: ["string", "null"]
        ataDate:
          type: ["string", "null"]
        ataTime:
          type: ["string", "null"]
        lkwArt:
          type: ["string", "null"]
          enum: [Container, Anlieferung, Abholung, Retoure, Sonstiges, null]
        status:
          type: string
          enum: [Angemeldet, Erwartet, Angekommen, Verladen, Abgefertigt, Storniert]
        tor:
          type: ["string", "null"]
    ShipmentUpdate:
      type: object
      properties:
        bezeichnung:
          type: ["string", "null"]
        kennzeichen:
          type: ["string", "null"]
        relation:
          type: ["string", "null"]
        speditionId:
          type: ["integer", "null"]
        subSpeditionId:
          type: ["integer", "null"]
        bemerkungen:
          type: ["string", "null"]
        telefon:
          type: ["string", "null"]
        etaDate:
          type: ["string", "null"]
        etaTime:
          type: ["string", "null"]
        ataDate:
          type: ["string", "null"]
        ataTime:
          type: ["string", "null"]
        lkwArt:
          type: ["string", "null"]
          enum: [Container, Anlieferung, Abholung, Retoure, Sonstiges, null]
        status:
          type: string
          enum: [Angemeldet, Erwartet, Angekommen, Verladen, Abgefertigt, Storniert]
        tor:
          type: ["string", "null"]
    BulkShipmentInput:
      type: object
      required: [shipments]
      properties:
        shipments:
          type: array
          items:
            $ref: "#/components/schemas/ShipmentInput"
...
    AuditEntry:
      type: object
      required: [id, module, recordId, changedAt]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        username:
          type: ["string", "null"]
        module:
          type: string
        recordId:
          type: integer
        field:
          type: ["string", "null"]
        oldValue:
          type: ["string", "null"]
        newValue:
          type: ["string", "null"]
        changedAt:
          type: string
...