# lib/api-zod/src/generated/types/auditEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * COMET LKW-Verladungsverwaltung API
 * OpenAPI spec version: 0.1.0
 */

export interface AuditEntry {
  id: number;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  username?: string | null;
  module: string;
  recordId: number;
  /** @nullable */
  field?: string | null;
  /** @nullable */
  oldValue?: string | null;
  /** @nullable */
  newValue?: string | null;
  changedAt: string;
}